update error handling

+++ app/write_row.py
@@ -24,7 +24,7 @@ def write_row(db_config, detection, camera):
       create table {camera} (
         filename   VARCHAR PRIMARY KEY NOT NULL,
         label      VARCHAR,
-        confidence REAL,
+        confidence INT,
         ymincoord       INT,
         ymaxcoord       INT,
         xmincoord       INT,
